graph refactor
